Small Side Project
Similar to the idea of QueenDroneCommunicationGP_1. I planned to make it
closer to real time. I changed from three floating tiles to an entire
floor of tiles. This removed the over the top environmental damage
effect. The tiles would fall away if the drone spent too much time on
top. I had hoped that communication would have more options with more
dimensions of movement. I had also hoped to make use of the time in
between of commands as part of the fitness function.